Lazy Composer MSBios Commit Script

+++ src/UseCase/Work/Member/Group/Remove/Handler.php
@@ -7,6 +7,7 @@
 namespace App\UseCase\Work\Member\Group\Remove;
 
 use App\Entity\Work\Member\Group;
+use App\Exception\DomainException;
 use Doctrine\ORM\EntityManagerInterface;
 
 /**
@@ -29,11 +30,17 @@ class Handler
 
     /**
      * @param Command $command
+     * @throws \Throwable
      */
     public function handle(Command $command): void
     {
         /** @var Group $group */
         $group = $this->em->find(Group::class, $command->id);
+
+        if ($group->getMembers()->count()) {
+            throw DomainException::groupIsNotEmpty();
+        }
+
         $this->em->remove($group);
         $this->em->flush();
     }
